Backend/requirements.txt
flask-cors==3.0.8
flask-jwt-extended==3.24.1
flask-migrate==2.5.3
flask-wtf==0.14.3
wtforms
flask_login
python-dotenv
